requirements.txt
langchain
langchain-google-genai
python-dateutil
pydantic
faiss-cpu
langchain-community
python-dotenv
langchain-core